refactor: moved ControlBar, SimpleVoiceAssistant, SuspectCard components into seperate file
