+++ frontend/app/interrogate/[suspectId]/page.tsx
@@ -1,22 +1,21 @@
 'use client';
 
 import { useParams, useRouter } from 'next/navigation';
-import { AnimatePresence, motion } from 'framer-motion';
+
 import {
 	LiveKitRoom,
-	useVoiceAssistant,
-	BarVisualizer,
+
+
 	RoomAudioRenderer,
-	VoiceAssistantControlBar,
+
 	AgentState,
-	DisconnectButton,
 } from '@livekit/components-react';
-import { useCallback, useEffect, useState } from 'react';
-import { MediaDeviceFailure } from 'livekit-client';
+import { useCallback, useState } from 'react';
 import type { ConnectionDetails } from '../../api/connection-details/route';
 import { NoAgentNotification } from '@/components/NoAgentNotification';
-import { CloseIcon } from '@/components/CloseIcon';
-import { useKrispNoiseFilter } from '@livekit/components-react/krisp';
+import { SimpleVoiceAssistant } from '@/components/interrogation/SimpleVoiceAssistant';
+import { ControlBar } from '@/components/interrogation/ControlBar';
+import { onDeviceFailure } from '@/utils/deviceFailure';
 
 /**
  * Interrogation Page Component
@@ -106,92 +105,5 @@ export default function InterrogationPage() {
 				</LiveKitRoom>
 			</main>
 		</div>
-	);
-}
-
-/**
- * SimpleVoiceAssistant Component
- * Displays the audio visualization and handles voice assistant state
- */
-function SimpleVoiceAssistant(props: {
-	onStateChange: (state: AgentState) => void;
-}) {
-	const { state, audioTrack } = useVoiceAssistant();
-
-	useEffect(() => {
-		props.onStateChange(state);
-	}, [props, state]);
-
-	return (
-		<div className="h-[300px] max-w-[90vw] mx-auto">
-			<BarVisualizer
-				state={state}
-				barCount={5}
-				trackRef={audioTrack}
-				className="agent-visualizer"
-				options={{ minHeight: 24 }}
-			/>
-		</div>
-	);
-}
-
-/**
- * ControlBar Component
- * Handles the UI controls for starting/stopping conversations
- */
-function ControlBar(props: {
-	onConnectButtonClicked: () => void;
-	agentState: AgentState;
-}) {
-	const krisp = useKrispNoiseFilter();
-	useEffect(() => {
-		krisp.setNoiseFilterEnabled(true);
-	}, [krisp]);
-
-	return (
-		<div className="relative h-[100px]">
-			<AnimatePresence>
-				{props.agentState === 'disconnected' && (
-					<motion.button
-						initial={{ opacity: 0, top: 0 }}
-						animate={{ opacity: 1 }}
-						exit={{ opacity: 0, top: '-10px' }}
-						transition={{ duration: 1, ease: [0.09, 1.04, 0.245, 1.055] }}
-						className="uppercase absolute left-1/2 -translate-x-1/2 px-4 py-2 bg-white text-black rounded-md"
-						onClick={() => props.onConnectButtonClicked()}
-					>
-						Start a conversation
-					</motion.button>
-				)}
-			</AnimatePresence>
-
-			<AnimatePresence>
-				{props.agentState !== 'disconnected' &&
-					props.agentState !== 'connecting' && (
-						<motion.div
-							initial={{ opacity: 0, top: '10px' }}
-							animate={{ opacity: 1, top: 0 }}
-							exit={{ opacity: 0, top: '-10px' }}
-							transition={{ duration: 0.4, ease: [0.09, 1.04, 0.245, 1.055] }}
-							className="flex h-8 absolute left-1/2 -translate-x-1/2 justify-center"
-						>
-							<VoiceAssistantControlBar controls={{ leave: false }} />
-							<DisconnectButton>
-								<CloseIcon />
-							</DisconnectButton>
-						</motion.div>
-					)}
-			</AnimatePresence>
-		</div>
-	);
-}
-
-/**
- * Error handler for media device failures
- */
-function onDeviceFailure(error?: MediaDeviceFailure) {
-	console.error(error);
-	alert(
-		'Error acquiring camera or microphone permissions. Please make sure you grant the necessary permissions in your browser and reload the tab'
 	);
 }
